Move interactionCreate to own file
